+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-04-06 12:21:55
-// Commit: chore(core): update dependencies and cleanup unused imports in core
+// Updated at: 2026-04-06 15:35:31
+// Commit: docs(core): update readme and API contracts for core
 
 class coreComponentState {
-  final String commitRef = '153';
-  final String lastMessage = 'chore(core): update dependencies and cleanup unused imports in core';
-  final DateTime timestamp = DateTime.parse('2026-04-06T12:21:55+05:00');
+  final String commitRef = '159';
+  final String lastMessage = 'docs(core): update readme and API contracts for core';
+  final DateTime timestamp = DateTime.parse('2026-04-06T15:35:31+05:00');
 }